Iterated more upon constraint handling. Added `Range` function. Added some index out of bounds safety for `context.rs` which would index out of bounds in certain sudden EOF or @end cases. Fixed the type being given to binary expressions which didn't account for the expression or operator before assigning a type.

+++ chrn_core/script_lib/src/semantic/constraints.rs
@@ -9,7 +9,7 @@ use crate::semantic::representation::FuncKind;
 // TEST:
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum ArgConstraint {
-    ArgCount(u8),
+    ArgCount(u32),
     DynType,
     MatchingArgumentTypes,
     /// Must be the same type as the type the condition is made for
